# api/http/v1/openapi.types.go
// Package v1 provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.4.1 DO NOT EDIT.
package v1

// Defines values for ClickActionAction.
const (
	OpenActivity ClickActionAction = "openActivity"
	OpenApp      ClickActionAction = "openApp"
	OpenUrl      ClickActionAction = "openUrl"
)

// Defines values for MessageRequestPriority.
const (
	High   MessageRequestPriority = "high"
	Low    MessageRequestPriority = "low"
	Normal MessageRequestPriority = "normal"
)

// ClickAction 点击动作，用于指定消息的点击行为。
type ClickAction struct {
	// Action 点击动作类型。
	Action ClickActionAction `json:"Action,omitempty"`

	// Activity 指定打开的 Activity 类名。
	Activity string `json:"Activity,omitempty"`

	// Parameters 附加参数，可以传递自定义数据。
	Parameters map[string]interface{} `json:"Parameters,omitempty"`

	// URL 指定打开的 URL 地址。
	URL string `json:"URL,omitempty"`
}

// ClickActionAction 点击动作类型。
type ClickActionAction string

// MessageRequest defines model for MessageRequest.
type MessageRequest struct {
	// TTL 消息的存活时间（以秒为单位）。
	TTL *int `json:"TTL,omitempty"`

	// Category 消息的类别，用于分类和识别消息类型。
	Category *string `json:"category,omitempty"`

	// ClickAction 点击动作，用于指定消息的点击行为。
	ClickAction *ClickAction `json:"clickAction,omitempty"`

	// Content 消息的主体内容。
	Content *string `json:"content,omitempty"`

	// Data 附加数据，可以包含额外的信息或操作。
	Data *map[string]interface{} `json:"data,omitempty"`

	// Foreground 指定消息是否在应用处于前台时显示。
	Foreground *bool `json:"foreground,omitempty"`

	// Icon 消息的图标，可以是图标的 URL 或者资源 ID。
	Icon *string `json:"icon,omitempty"`

	// NotifyType 通知的类型，比如“消息通知”，“系统通知”等。
	NotifyType *int `json:"notifyType,omitempty"`

	// Priority 消息的优先级。
	Priority *MessageRequestPriority `json:"priority,omitempty"`

	// Sound 消息的提示音。
	Sound *string `json:"sound,omitempty"`

	// Title 消息标题。
	Title *string `json:"title,omitempty"`
}

// MessageRequestPriority 消息的优先级。
type MessageRequestPriority string

// Response defines model for Response.
type Response struct {
	// Code 响应状态码。
	Code int `json:"code,omitempty"`

	// Data 响应的具体数据，可能为空或包含详细信息。
	Data map[string]interface{} `json:"data,omitempty"`

	// Message 响应消息，提供状态说明。
	Message string `json:"message,omitempty"`
}

// PushMessageParams defines parameters for PushMessage.
type PushMessageParams struct {
	// DryRun 是否为 DryRun 模式，如果为 `true`，仅校验配置和构建通知信息不会进行实际推送。
	DryRun *bool `form:"dryRun,omitempty" json:"dryRun,omitempty"`

	// Retry 最大重试次数，如果推送失败，将根据该次数进行重试。默认为 0 表示不重试。
	Retry *int `form:"retry,omitempty" json:"retry,omitempty"`

	// RetryInterval 每次重试之间的时间间隔（单位：秒），仅在 `retry` 大于 0 时有效。
	RetryInterval *int `form:"retryInterval,omitempty" json:"retryInterval,omitempty"`
}

// PushMessageJSONRequestBody defines body for PushMessage for application/json ContentType.
type PushMessageJSONRequestBody = MessageRequest